# app/Http/Controllers/GameController.php
<?php

namespace App\Http\Controllers;

use App\Actions\Games\JoinGameAction;
use App\Actions\Games\LeaveGameAction;
use App\Exceptions\Games\GameIsFullException;
use App\Models\Game;
use App\Models\Player;
use Illuminate\Http\Request;

/**
 * @OA\Tag(
 *     name="Games",
 *     description="Gerenciamento de partidas"
 * )
 */
class GameController extends Controller
{
    /**
     * @OA\Get(
     *     path="/api/game",
     *     tags={"Games"},
     *     summary="Lista partidas do usuário",
     *     description="Lista as partidas do usuário logado (participando ou convidado) com filtros opcionais. Se status não for informado, retorna apenas partidas com status 'open'.",
     *     security={{"bearerAuth":{}}},
     *
     *     @OA\Parameter(
     *         name="data_time",
     *         in="query",
     *         required=false,
     *         description="Filtrar por data da partida (formato: YYYY-MM-DD)",
     *         @OA\Schema(type="string", format="date", example="2026-02-10")
     *     ),
     *     @OA\Parameter(
     *         name="club_id",
     *         in="query",
     *         required=false,
     *         description="Filtrar por ID do clube",
     *         @OA\Schema(type="integer", example=1)
     *     ),
     *     @OA\Parameter(
     *         name="min_level",
     *         in="query",
     *         required=false,
     *         description="Filtrar por nível mínimo da partida",
     *         @OA\Schema(type="integer", example=2)
     *     ),
     *     @OA\Parameter(
     *         name="max_level",
     *         in="query",
     *         required=false,
     *         description="Filtrar por nível máximo da partida",
     *         @OA\Schema(type="integer", example=4)
     *     ),
     *     @OA\Parameter(
     *         name="status",
     *         in="query",
     *         required=false,
     *         description="Filtrar por status da partida (padrão: open)",
     *         @OA\Schema(type="string", enum={"open","full","in_progress","completed","canceled"}, example="open")
     *     ),
     *
     *     @OA\Response(
     *         response=200,
     *         description="Dados da partida e players daquela partida",
     *         @OA\JsonContent(
     *             type="array",
     *             @OA\Items(type="object")
     *         )
     *     )
     * )
     */
    public function index(Request $request)
    {
        $player = $request->user()->player;

        if (!$player) {
            return response()->json([
                'message' => 'Usuário não possui player vinculado'
            ], 400);
        }

        $status = $request->query('status', 'open');
        $eagerLoad = [
            'players:id,full_name,level,side,profile_image_url',
            'owner:id,full_name',
            'club:id,name,city,state',
            'court:id,club_id,name,type,covered,price_per_hour'
        ];

        // Partidas que o player já participa
        $joinedGames = $player->games()
            ->where('status', $status)
            ->when($request->query('data_time'), fn ($q, $date) => $q->whereDate('data_time', $date))
            ->when($request->query('club_id'), fn ($q, $clubId) => $q->where('club_id', $clubId))
            ->when($request->query('min_level'), fn ($q, $level) => $q->where('min_level', '>=', $level))
            ->when($request->query('max_level'), fn ($q, $level) => $q->where('max_level', '<=', $level))
            ->with($eagerLoad)
            ->get();

        // Partidas privadas onde tem convite aceito mas ainda nao entrou
        $invitedGameIds = $player->gameInvitations()
            ->where('status', 'accepted')
            ->pluck('game_id');

        $invitedGames = Game::whereIn('id', $invitedGameIds)
            ->where('status', $status)
            ->where('type', 'private')
            ->whereDoesntHave('players', fn ($q) => $q->where('players.id', $player->id))
            ->when($request->query('data_time'), fn ($q, $date) => $q->whereDate('data_time', $date))
            ->when($request->query('club_id'), fn ($q, $clubId) => $q->where('club_id', $clubId))
            ->when($request->query('min_level'), fn ($q, $level) => $q->where('min_level', '>=', $level))
            ->when($request->query('max_level'), fn ($q, $level) => $q->where('max_level', '<=', $level))
            ->with($eagerLoad)
            ->get();

        $games = $joinedGames->merge($invitedGames)->unique('id')->values();

        $games->each(function ($game) {
            $game->makeHidden(['created_at', 'updated_at']);
            $game->players->each(function ($player) {
                $player->pivot->makeHidden(['created_at', 'updated_at']);
            });
        });

        return response()->json($games);
    }

    /**
     * @OA\Get(
     *     path="/api/game/available",
     *     tags={"Games"},
     *     summary="Lista partidas disponíveis",
     *     description="Retorna partidas públicas com status open e partidas privadas onde o jogador foi convidado",
     *     security={{"bearerAuth":{}}},
     *
     *     @OA\Response(
     *         response=200,
     *         description="Lista de partidas públicas disponíveis",
     *         @OA\JsonContent(
     *             type="array",
     *             @OA\Items(type="object")
     *         )
     *     ),
     *
     *     @OA\Response(
     *         response=400,
     *         description="Usuário não possui player vinculado"
     *     )
     * )
     */
    public function available(Request $request)
    {
        $player = $request->user()->player;

        if (!$player) {
            return response()->json([
                'message' => 'Usuário não possui player vinculado'
            ], 400);
        }

        $eagerLoad = [
            'players:id,full_name,level,side,profile_image_url',
            'owner:id,full_name',
            'club:id,name,city,state',
            'court:id,club_id,name,type,covered,price_per_hour'
        ];

        // Partidas públicas abertas (exceto as do próprio player)
        $publicGames = Game::where('type', 'public')
            ->where('status', 'open')
            ->where('owner_player_id', '!=', $player->id)
            ->with($eagerLoad)
            ->get();

        // Partidas privadas onde o player foi convidado (pending ou accepted)
        $invitedGameIds = $player->gameInvitations()
            ->whereIn('status', ['pending', 'accepted'])
            ->pluck('game_id');

        $invitedGames = Game::whereIn('id', $invitedGameIds)
            ->where('status', 'open')
            ->where('type', 'private')
            ->with($eagerLoad)
            ->get();

        $games = $publicGames->merge($invitedGames)->unique('id')->values();

        $games->each(function ($game) {
            $game->makeHidden(['created_at', 'updated_at']);
            $game->players->each(function ($player) {
                $player->pivot->makeHidden(['created_at', 'updated_at']);
            });
        });

        return response()->json($games);
    }

    /**
     * @OA\Get(
     *     path="/api/game/id",
     *     tags={"Games"},
     *     summary="Lista uma partida",
     *     description="Lista uma partida conforme id solicitado. Partidas privadas são visíveis apenas para o proprietário e jogadores convidados.",
     *     security={{"bearerAuth":{}}},
     *
     *     @OA\Response(
     *         response=200,
     *         description="Lista a partida",
     *         @OA\JsonContent(
     *             type="array",
     *             @OA\Items(type="object")
     *         )
     *     ),
     *     @OA\Response(
     *         response=403,
     *         description="Sem permissão para visualizar esta partida"
     *     )
     * )
     */
    public function show(Request $request, $id)
    {
        $game = Game::with([
            'players:id,full_name,level,side,profile_image_url',
            'owner:id,full_name',
            'club:id,name,city,state',
            'court:id,club_id,name,type,covered,price_per_hour'
        ])->findOrFail($id);

        if ($request->user()->cannot('view', $game)) {
            return response()->json([
                'message' => 'Sem permissão para visualizar esta partida'
            ], 403);
        }

        $game->makeHidden(['created_at', 'updated_at']);
        $game->players->each(function ($player) {
            $player->pivot->makeHidden(['created_at', 'updated_at']);
        });

        return response()->json($game);
    }

    /**
     * @OA\Post(
     *     path="/api/game",
     *     tags={"Games"},
     *     summary="Cria uma nova partida",
     *     description="Cria uma nova partida associada ao usuário autenticado",
     *     security={{"bearerAuth":{}}},
     *
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             required={"type","data_time","club_id","court_id","status","game_type"},
     *             @OA\Property(property="title", type="string", example="Partida de Padel"),
     *             @OA\Property(property="description", type="string", example="Partida amistosa"),
     *             @OA\Property(property="type", type="string", enum={"public","private"}, example="public"),
     *             @OA\Property(property="data_time", type="string", format="date-time", example="2026-02-10 19:00:00"),
     *             @OA\Property(property="club_id", type="integer", example=1),
     *             @OA\Property(property="court_id", type="integer", example=3),
     *             @OA\Property(property="custom_location", type="string", example="Quadra externa"),
     *             @OA\Property(property="min_level", type="integer", example=2),
     *             @OA\Property(property="max_level", type="integer", example=4),
     *             @OA\Property(property="max_players", type="integer", example=4),
     *             @OA\Property(property="status", type="string", enum={"open","full","in_progress","completed","canceled"}, example="open"),
     *             @OA\Property(property="price", type="number", format="float", example=120),
     *             @OA\Property(property="cost_per_player", type="number", format="float", example=30),
     *             @OA\Property(property="game_type", type="string", enum={"casual","competitive","training"}, example="casual"),
     *             @OA\Property(property="duration_minutes", type="integer", example=90)
     *         )
     *     ),
     *
     *     @OA\Response(
     *         response=201,
     *         description="Partida criada com sucesso",
     *         @OA\JsonContent(type="object")
     *     ),
     *
     *     @OA\Response(
     *         response=422,
     *         description="Erro de validação"
     *     )
     * )
     */
    public function store(Request $request)
    {
        $data = $request->validate([
            "title" => 'nullable|string|max:255',
            "description" => 'nullable|string|max:500',
            "type"  => 'required|in:public,private',
            "data_time" => 'nullable|date',
            "club_id" => 'required|exists:clubs,id',
            "court_id" => 'required|exists:courts,id',
            "custom_location" => 'nullable|string|max:500',
            "min_level" => 'nullable|integer',
            "max_level" => 'nullable|integer',
            "max_players" => 'nullable|integer|min:2',
            "price" => 'nullable|numeric',
            "cost_per_player" => 'nullable|numeric',
            "game_type" => 'required|in:casual,competitive,training',
            "duration_minutes" => 'nullable|integer|min:30'
        ]);

        $player = $request->user()->player;

        if (!$player) {
            return response()->json([
                'message' => 'Usuário não possui player vinculado'
            ], 400);
        }

        $game = new Game($data);
        $game->owner_player_id = $player->id;
        $game->status = 'open';
        $game->save();

        $game->players()->attach($player->id, [
            'joined_at' => now()
        ]);

        return response()->json(
            $game->load('players'),
            201
        );
    }

    /**
     * @OA\Post(
     *     path="/api/game/{game}/join",
     *     tags={"Games"},
     *     summary="Entrar em uma partida",
     *     description="Adiciona o player do usuário autenticado a uma partida. Partidas públicas são abertas a todos. Partidas privadas requerem convite aceito.",
     *     security={{"bearerAuth":{}}},
     *
     *     @OA\Parameter(
     *         name="game",
     *         in="path",
     *         required=true,
     *         description="ID da partida",
     *         @OA\Schema(type="integer")
     *     ),
     *
     *     @OA\Response(
     *         response=200,
     *         description="Entrou na partida com sucesso",
     *         @OA\JsonContent(
     *             type="object",
     *             @OA\Property(property="message", type="string", example="Entrou na partida")
     *         )
     *     ),
     *
     *     @OA\Response(
     *         response=400,
     *         description="Usuário não possui player vinculado"
     *     ),
     *     @OA\Response(
     *         response=403,
     *         description="Partida privada - acesso não permitido"
     *     ),
     *     @OA\Response(
     *         response=404,
     *         description="Partida não encontrada"
     *     ),
     *     @OA\Response(
     *         response=409,
     *         description="Conflito - partida fechada ou cheia"
     *     )
     * )
     */
    public function join(Request $request, Game $game)
    {
        $player = $request->user()->player;

        if (!$player) {
            return response()->json(['message' => 'Usuário não possui player vinculado'], 400);
        }

        if ($request->user()->cannot('join', $game)) {
            return response()->json(['message' => 'Sem permissão para entrar nesta partida'], 403);
        }

        try {
            app(JoinGameAction::class)->execute($game, $player);
        } catch (GameIsFullException $e) {
            return response()->json(['message' => $e->getMessage()], 409);
        }

        return response()->json(['message' => 'Entrou na partida']);
    }

    /**
     * @OA\Post(
     *     path="/api/game/{game}/leave",
     *     tags={"Games"},
     *     summary="Sair de uma partida",
     *     description="Remove o player do usuário autenticado de uma partida existente. O criador da partida não pode sair.",
     *     security={{"bearerAuth":{}}},
     *
     *     @OA\Parameter(
     *         name="game",
     *         in="path",
     *         required=true,
     *         description="ID da partida",
     *         @OA\Schema(type="integer")
     *     ),
     *
     *     @OA\Response(
     *         response=200,
     *         description="Saiu da partida com sucesso",
     *         @OA\JsonContent(
     *             type="object",
     *             @OA\Property(property="message", type="string", example="Saiu da partida")
     *         )
     *     ),
     *
     *     @OA\Response(
     *         response=400,
     *         description="Usuário não possui player vinculado"
     *     ),
     *     @OA\Response(
     *         response=404,
     *         description="Partida não encontrada"
     *     ),
     *     @OA\Response(
     *         response=409,
     *         description="Conflito - criador não pode sair da partida"
     *     )
     * )
     */
    public function leave(Request $request, Game $game)
    {
        $player = $request->user()->player;

        if (!$player) {
            return response()->json(['message' => 'Usuário não possui player vinculado'], 400);
        }

        if ($game->owner_player_id === $player->id) {
            return response()->json(['message' => 'O criador da partida não pode sair. Cancele a partida.'], 409);
        }

        app(LeaveGameAction::class)->execute($game, $player);

        return response()->json(['message' => 'Saiu da partida']);
    }

    /**
     * @OA\Delete(
     *     path="/api/game/{game}/players/{player}",
     *     tags={"Games"},
     *     summary="Remove um jogador da partida",
     *     description="Permite ao dono da partida remover outro jogador. O dono não pode remover a si mesmo.",
     *     security={{"bearerAuth":{}}},
     *
     *     @OA\Parameter(
     *         name="game",
     *         in="path",
     *         required=true,
     *         description="ID da partida",
     *         @OA\Schema(type="integer")
     *     ),
     *     @OA\Parameter(
     *         name="player",
     *         in="path",
     *         required=true,
     *         description="ID do jogador a ser removido",
     *         @OA\Schema(type="integer")
     *     ),
     *
     *     @OA\Response(
     *         response=200,
     *         description="Jogador removido da partida",
     *         @OA\JsonContent(
     *             @OA\Property(property="message", type="string", example="Jogador removido da partida")
     *         )
     *     ),
     *     @OA\Response(response=400, description="Usuário não possui player vinculado"),
     *     @OA\Response(response=403, description="Apenas o dono da partida pode remover jogadores"),
     *     @OA\Response(response=404, description="Jogador não está na partida"),
     *     @OA\Response(response=409, description="O dono não pode remover a si mesmo")
     * )
     */
    public function removePlayer(Request $request, Game $game, Player $player)
    {
        $currentPlayer = $request->user()->player;

        if (!$currentPlayer) {
            return response()->json(['message' => 'Usuário não possui player vinculado'], 400);
        }

        if ($game->owner_player_id !== $currentPlayer->id) {
            return response()->json(['message' => 'Apenas o dono da partida pode remover jogadores'], 403);
        }

        if ($player->id === $currentPlayer->id) {
            return response()->json(['message' => 'O dono não pode remover a si mesmo da partida'], 409);
        }

        $isInGame = $game->players()->where('players.id', $player->id)->exists();

        if (!$isInGame) {
            return response()->json(['message' => 'Jogador não está na partida'], 404);
        }

        $game->players()->detach($player->id);

        if ($game->status === 'full') {
            $game->update(['status' => 'open']);
        }

        return response()->json(['message' => 'Jogador removido da partida'], 200);
    }

    /**
     * @OA\Put(
     *     path="/api/game/{id}",
     *     tags={"Games"},
     *     summary="Atualiza uma partida",
     *     description="Atualiza os dados de uma partida existente",
     *     security={{"bearerAuth":{}}},
     *
     *     @OA\Parameter(
     *         name="id",
     *         in="path",
     *         required=true,
     *         description="ID da partida",
     *         @OA\Schema(type="integer")
     *     ),
     *
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(type="object")
     *     ),
     *
     *     @OA\Response(
     *         response=200,
     *         description="Clube atualizado",
     *         @OA\JsonContent(type="object")
     *     ),
     *
     *     @OA\Response(
     *         response=404,
     *         description="Partida não encontrada"
     *     )
     * )
     */
    public function update(Request $request, $id)
    {
        $game = Game::find($id);

        if (!$game) {
            return response()->json([
                'message' => 'Partida não encontrada'
            ], 404);
        }

        $data = $request->validate([
            "title" => 'nullable|string|max:255',
            "description" => 'nullable|string|max:500',
            "type"  => 'required|in:public,private',
            "data_time" => 'nullable|date',
            "club_id" => 'required',
            "court_id" => 'required',
            "custom_location" => 'nullable|string|max:500',
            "min_level" => 'nullable',
            "max_level" => 'nullable',
            "max_players" => 'nullable',
            "status"     => 'required|in:open,full,in_progress,completed,canceled',
            "price" => 'nullable',
            "cost_per_player" => 'nullable',
            "game_type" => 'required|in:casual,competitive,training',
            "duration_minutes" => 'nullable'
        ], [
            'type.required' => 'O tipo de partida é obrigatório. Defina entre publica ou privada',
            'status.required' => 'A status da partida é obrigatório.',
            'data_time.required' => 'A data e hora da partida é obrigatório.',
            'club_id.required' => 'O clube é obrigatório.',
            'court_id.required' => 'A quadra é obrigatório.',
            'court_id.required' => 'O clube é obrigatório.',
        ]);

        $game->update($data);

        return response()->json($game);
    }
}