# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Padel Match API",
        "description": "Documentação da API do Padel Match",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/game": {
            "get": {
                "tags": [
                    "Games"
                ],
                "summary": "Lista partidas do usuário",
                "description": "Lista as partidas do usuário logado (participando ou convidado) com filtros opcionais. Se status não for informado, retorna apenas partidas com status 'open'.",
                "operationId": "589589f26003a22c288f69fb4ded72b0",
                "parameters": [
                    {
                        "name": "data_time",
                        "in": "query",
                        "description": "Filtrar por data da partida (formato: YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-02-10"
                        }
                    },
                    {
                        "name": "club_id",
                        "in": "query",
                        "description": "Filtrar por ID do clube",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "min_level",
                        "in": "query",
                        "description": "Filtrar por nível mínimo da partida",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    },
                    {
                        "name": "max_level",
                        "in": "query",
                        "description": "Filtrar por nível máximo da partida",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 4
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por status da partida (padrão: open)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "open",
                                "full",
                                "in_progress",
                                "completed",
                                "canceled"
                            ],
                            "example": "open"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados da partida e players daquela partida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Games"
                ],
                "summary": "Cria uma nova partida",
                "description": "Cria uma nova partida associada ao usuário autenticado",
                "operationId": "a0adcc6e849923fc7808a9337b1cb4ce",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "data_time",
                                    "club_id",
                                    "court_id",
                                    "status",
                                    "game_type"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Partida de Padel"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Partida amistosa"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "public",
                                            "private"
                                        ],
                                        "example": "public"
                                    },
                                    "data_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-02-10 19:00:00"
                                    },
                                    "club_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "court_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "custom_location": {
                                        "type": "string",
                                        "example": "Quadra externa"
                                    },
                                    "min_level": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "max_level": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "max_players": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "open",
                                            "full",
                                            "in_progress",
                                            "completed",
                                            "canceled"
                                        ],
                                        "example": "open"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 120
                                    },
                                    "cost_per_player": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 30
                                    },
                                    "game_type": {
                                        "type": "string",
                                        "enum": [
                                            "casual",
                                            "competitive",
                                            "training"
                                        ],
                                        "example": "casual"
                                    },
                                    "duration_minutes": {
                                        "type": "integer",
                                        "example": 90
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Partida criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/game/available": {
            "get": {
                "tags": [
                    "Games"
                ],
                "summary": "Lista partidas disponíveis",
                "description": "Retorna partidas públicas com status open e partidas privadas onde o jogador foi convidado",
                "operationId": "72be6a94a6467ddf1870ca68fa973f57",
                "responses": {
                    "200": {
                        "description": "Lista de partidas públicas disponíveis",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Usuário não possui player vinculado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/game/id": {
            "get": {
                "tags": [
                    "Games"
                ],
                "summary": "Lista uma partida",
                "description": "Lista uma partida conforme id solicitado. Partidas privadas são visíveis apenas para o proprietário e jogadores convidados.",
                "operationId": "a97bf63cceaee3b8908a3d1854eddf6f",
                "responses": {
                    "200": {
                        "description": "Lista a partida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sem permissão para visualizar esta partida"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}